added mark enum and refactored

+++ src/main/java/pl/schoolmanagementsystem/common/mark/Mark.java
@@ -1,10 +1,6 @@
 package pl.schoolmanagementsystem.common.mark;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -23,7 +19,7 @@ public class Mark {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private byte mark;
+    private double mark;
 
     private long studentId;
